Develop

+++ src/main/java/com/myproject/restexample/repository/StudentRepository.java
@@ -1,11 +1,14 @@
 package com.myproject.restexample.repository;
 
+
 import org.springframework.data.jpa.repository.JpaRepository;
 
 import com.myproject.restexample.entity.Student;
 
 
-public interface StudentRepository extends JpaRepository<Student, Long>{
+public interface StudentRepository extends JpaRepository<Student, Integer>{
 
-	Student getById(Long id);
+	//Student getById(Long id);
+	Student getById(Integer id);
+	
 }
